update(scripts): use Java 17 for all jars and allow skipping fix

+++ Darcy/runAll.sh
@@ -1,10 +1,21 @@
 #!/bin/bash
 
+# script usage is ./runAll.sh <projectPath> [-fix]
+
+
 # check if argument is provided
 if [ -z "$1" ]
   then
-    >&2 echo "No project path argument supplied"
+    >&2 echo "Usage: ./runAll.sh <projectPath>"
     exit 1
+fi
+
+# check if flag named -fix is set
+if [[ $* == *--fix* ]]
+  then
+    fix=true
+  else
+    fix=false
 fi
 
 projectPath=$1
@@ -15,11 +26,14 @@ if [ ! -d "$projectPath" ]
     exit 1
 fi
 
-JAVA_8_HOME=/usr/lib/jvm/amazon-corretto-8.372.07.1-linux-x64
-JAVA_17_HOME=/usr/lib/jvm/jdk-17
+JAVA_HOME=/usr/lib/jvm/jdk-17
 
 foo=" "
-rm -r ./XMLreports/
+# check if XMLreports directory exists
+if [ -d "./XMLreports" ]
+  then
+    rm -r ./XMLreports/
+fi
 mkdir XMLreports
 
 echo "======== Constructing modules.txt and module-infos.txt ========"
@@ -64,38 +78,30 @@ echo "-------- Finished Checking Inconsistencies --------"
 
 echo "======== Checking for opens Inconsistencies: Java Reflection Analysis ========"
 start_open=$(ruby -e 'puts (Time.now.to_f * 1000).to_i')
-# changing Java to 1.8
-#export JAVA_HOME=/Library/Java/JavaVirtualMachines/jdk1.8.0_101.jdk/Contents/Home/
-# sudo update-alternatives --set java $JAVA_8_HOME/bin/java
-# sudo update-alternatives --set javac $JAVA_8_HOME/bin/javac
-
-$JAVA_8_HOME/bin/java -jar java-reflection-analysis.jar
+$JAVA_HOME/bin/java -jar java-reflection-analysis.jar
 end_open=$(ruby -e 'puts (Time.now.to_f * 1000).to_i')
 elapsed_open=$((end_open - start_open))
 echo "-------- Finished Checking opens Inconsistencies --------"
 
 echo "======== Checking for uses Inconsistencies: Uses Provides Analysis ========"
 start_uses=$(ruby -e 'puts (Time.now.to_f * 1000).to_i')
-
-java -jar uses-provides-analysis.jar
-
+$JAVA_HOME/bin/java -jar uses-provides-analysis.jar
 end_uses=$(ruby -e 'puts (Time.now.to_f * 1000).to_i')
 elapsed_uses=$((end_uses - start_uses))
 echo "-------- Finished Checking uses Inconsistencies --------"
 
-
 echo "======== Repairing: Transforming module-info Files ========"
 start_repair=$(ruby -e 'puts (Time.now.to_f * 1000).to_i')
-
-# changing Java to 1.9
-#export JAVA_HOME=/Library/Java/JavaVirtualMachines/jdk-9.jdk/Contents/Home/
-# sudo update-alternatives --set java $JAVA_17_HOME/bin/java
-# sudo update-alternatives --set javac $JAVA_17_HOME/bin/javac
-
-$JAVA_17_HOME/bin/java -jar ModuleInfoTransformer.jar
+if [ "$fix" = true ]
+  then
+    $JAVA_HOME/bin/java -jar ModuleInfoTransformer.jar
+  else
+    echo "Not fixing inconsistencies, skipping repair"
+fi
 end_repair=$(ruby -e 'puts (Time.now.to_f * 1000).to_i')
 elapsed_repair=$((end_repair - start_repair))
 echo "-------- Finished Repairing --------"
+
 echo "======== Stats ========"
 end_glob=$(ruby -e 'puts (Time.now.to_f * 1000).to_i')
 elapsed_total=$((end_glob - start_glob))
